Cool off after daemon crash; simplify startup
If the Erlang ssh daemon crashes, let them things cool off before
restarting. This is a very rare occurance that is observed, but not
reproducible via unit tests except by manually exiting the daemon.

The theory is that the previously aggressive restart strategy required
that sockets be forcefully closed, and that waiting would have been
sufficient.