Refactor to enable start non-app supervision tree
This moves quite a bit of code around, but the change is pretty small
since the application environment was previously being passed as
arguments to the daemon GenServer anyway. Most of the change was adding
the `case` to automatically start the daemon if `:nerves_ssh` was in the
application environment. If it's not, then nothing happens. This seems
quite reasonable since if you didn't specify anything in the application
config previously, the SSH daemon would have started, but you couldn't
log in.

The rest of the changes are adjusting unit tests and documentation.

Co-authored-by: Jon Carstens <jjcarstens@me.com>